select lowest rtt proxy when proxying
instead of distributing proxied sends over all available connected
proxy peers, select the one with the lowest round trip time.

RTT is calculated as an average of last 5 pings tracked between the
peers.

The RTT for peer A is just assumed to be the RTT to the proxy (which is
definitely incorrect as it could be more than double that, but at least indicative)