Game would hang on loader after reset
With two windows open, the database on the second window would get
cleared after the window had already partially initiallised. This would
cause a problem where settings would get created by the effect, the
settings being cleared and the settings effect not rerunning as the 'db'
dependency on the effect wouldn't change. By adding 'settings' as a
dependency it will trigger the effect to recreate them if they are
cleared

After fixing this the was the same problem with the client provider. The
client would be created and initialised which causes the network
settings to be stored in the db. The db is cleared after the client has
initialised where we are in a state where we have an initialised client
but no network settings and this would cause the game to hang at
'generating client keypair' in the terminal flow as it will wait
indefinitely for the network settings to be created. I solved this by
adding an effect that checks if the client has been initialised but with
missing network settings. I increment 'resetCount' which causes the
client to be shutdown and restarted. If this is too heavy handed I think
I could make the effect call 'init' on the client again but I wasn't
sure if it would play nicely being initialised twice