support pwa build
pull out all the socket-specific bits into a runtime:platform adapter
that can be either stubbed out or replaced with non-socket specific
parts.

this allows us to swap out the socket dgram implementation with one that
can work in the web browser context .... @playmint/wgram is a little package that
proxies dgram over websockets for this purpose.